CORE-9596 Some tidying
